Revert "Test Delete and Add functions"
