+++ src/main-src.js
@@ -3,11 +3,11 @@
 import './style-src.css';
 import Sortable from 'sortablejs';
 import {
-  showList, updateStorage, addTask, editTask, updateStatus, orderTasks, removeTask,
+  showList, updateStorage, addTask, editTask, updateStatus,
 } from './task.js';
 
 import {
-  getStorage, updateAfterDrag, removeCompletedTasks,
+  getStorage, updateAfterDrag, removeCompletedTasks, removeTask,
 } from './storage.js';
 
 let tasks = getStorage();
@@ -33,9 +33,6 @@ listDiv.addEventListener('click', (event) => {
     if (event.target.className === 'fas fa-trash') {
       tasks = getStorage();
       tasks = removeTask(tasks, parseInt(event.target.parentElement.id, 10));
-      orderTasks(tasks);
-      updateStorage(tasks);
-      showList(tasks);
     }
     if (event.target.classList.contains('description')) {
       tasks = getStorage();
@@ -47,14 +44,11 @@ listDiv.addEventListener('click', (event) => {
 
 clearAllButton.addEventListener('click', () => {
   tasks = removeCompletedTasks(tasks);
-  showList(tasks);
 });
 
 const addTaskButton = document.getElementById('add');
 addTaskButton.addEventListener('click', () => {
-  const description = document.getElementById('task-desc').value;
-  addTask(tasks, description);
-  document.getElementById('task-desc').value = '';
+  addTask(tasks);
   updateStorage(tasks);
   showList(tasks);
 });
@@ -65,7 +59,6 @@ new Sortable(listDiv, {
 });
 
 listDiv.addEventListener('dragend', (event) => {
-  tasks = updateAfterDrag(listDiv);
-  showList(tasks);
+  updateAfterDrag(listDiv);
   event.stopPropagation();
 });
