+++ src/task.js
@@ -43,15 +43,12 @@ export const updateStorage = (storage) => {
   localStorage.setItem('storage', JSON.stringify(storage));
 };
 
-export const addTask = (tasks, description) => {
+export const addTask = (tasks) => {
+  const description = document.getElementById('task-desc').value;
   const index = (tasks.length === 0) ? 1 : tasks[tasks.length - 1].index + 1;
   const completed = false;
   tasks.push({ description, index, completed });
-};
-
-export const removeTask = (tasks, id) => {
-  tasks = tasks.filter((task) => task.index !== id);
-  return tasks;
+  document.getElementById('task-desc').value = '';
 };
 
 export const editTask = (tasks, target) => {
